requirements.txt
Streamlit==0.84.0
pillow==8.3.1
scikit-learn==0.24.2